Moved augmentation functions to a new python file. This needs to run in its own environment which is not compatible with the driver's environment requirements.
